Calculate Average Distress score and track score based on session.
In the current session, N (I have considered 5) messages are used to calculate avg distress score so that it is considred to stop the conversation if the avg score is greater than 80.

In frontend, the avg distress score is displayed for each message so that we can see the differences based on each message.